+++ src/test/java/domain/card/DeckTest.java
@@ -1,7 +1,5 @@
-package domain;
+package domain.card;
 
-import domain.card.Deck;
-import domain.card.RandomCardGenerator;
 import org.assertj.core.api.InstanceOfAssertFactories;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
